get tmux

+++ install.sh
@@ -8,9 +8,27 @@ if ! command -v python3 >/dev/null 2>&1; then
     exit 1
 fi
 
+if ! command -v tmux >/dev/null 2>&1; then
+    read -p "tmux, a dependancy, is not installed. Would you like to install it now? (y/n) " answer
+    case "$answer" in
+        [Yy]* )
+             sudo apt-get install -y tmux
+            ;;
+        [Nn]* )
+            echo "Aborted."
+            exit 1
+            ;;
+        * )
+            exit 1
+            ;;
+    esac
+fi
+
 wget https://www.minecraft.net/bedrockdedicatedserver/bin-linux/bedrock-server-1.26.40.8.zip -O bedrock-server.zip
 
-unzip bedrock-server.zip -d bedrock-server
+echo "unzipping bedrock server..."
+unzip -q bedrock-server.zip -d bedrock-server
+chmod +x bedrock-server/bedrock_server
 
 rm bedrock-server.zip
